+++ scripts/release/changelog.js
@@ -7,15 +7,20 @@
 
 'use strict';
 
-require( '@ckeditor/ckeditor5-dev-release-tools' )
-	.generateChangelogForMonoRepository( {
-		cwd: process.cwd(),
-		packages: 'packages',
-		transformScope: name => {
-			if ( name === 'generator' ) {
-				return 'https://www.npmjs.com/package/ckeditor5-package-generator';
-			}
+const { generateChangelogForMonoRepository } = require( '@ckeditor/ckeditor5-dev-release-tools' );
+const parseArguments = require( './utils/parsearguments' );
 
-			return 'https://www.npmjs.com/package/@ckeditor/ckeditor5-package-' + name;
+const cliArguments = parseArguments( process.argv.slice( 2 ) );
+
+generateChangelogForMonoRepository( {
+	cwd: process.cwd(),
+	packages: 'packages',
+	releaseBranch: cliArguments.branch,
+	transformScope: name => {
+		if ( name === 'generator' ) {
+			return 'https://www.npmjs.com/package/ckeditor5-package-generator';
 		}
-	} );
+
+		return 'https://www.npmjs.com/package/@ckeditor/ckeditor5-package-' + name;
+	}
+} );